+++ src/App.jsx
@@ -1,29 +1,6 @@
 import React from 'react';
 import './App.scss';
-import { Person } from './components/Person/Person';
-
-export const misha = {
-  name: 'Misha',
-  age: 37,
-  sex: 'm',
-  isMarried: true,
-  partnerName: 'Natasha',
-};
-
-export const olya = {
-  name: 'Olya',
-  age: null,
-  sex: 'f',
-  isMarried: true,
-  partnerName: 'Maksym',
-};
-
-export const alex = {
-  name: 'Alex',
-  age: 25,
-  sex: 'm',
-  isMarried: false,
-};
+import { alex, misha, olya, Person } from './components/Person/Person';
 
 export const App = () => (
   <div className="App">
