+++ src/App.scss
@@ -14,7 +14,7 @@ iframe {
   }
 
   &__age {
-    font-weight: bold;
     color: #00f;
+    font-weight: bold;
   }
 }
