+++ src/components/Person/Person.jsx
@@ -1,11 +1,39 @@
-export const Person = ({ person }) => (
-  <section className="Person">
-    <h2 className="Person__name">My name is {person.name}</h2>
-    {person.age && <p className="Person__age">I am {person.age}</p>}
-    <p className="Person__partner">
-      {person.isMarried
-        ? `${person.partnerName} is my ${person.sex === 'm' ? 'wife' : 'husband'}`
-        : 'I am not married'}
-    </p>
-  </section>
-);
+export const Person = ({ person }) => {
+  const { name, age, sex, isMarried, partnerName } = person;
+
+  const partnerMass = sex === 'm' ? 'wife' : 'husband';
+
+  const isMarrieds = isMarried
+    ? `${partnerName} is my ${partnerMass}`
+    : 'I am not married';
+
+  return (
+    <section className="Person">
+      <h2 className="Person__name">My name is {name}</h2>
+      {age && <p className="Person_age">I am {age}</p>}
+      <p className="Person__partner">{isMarrieds}</p>
+    </section>
+  );
+};
+
+export const misha = {
+  name: 'Misha',
+  age: 37,
+  sex: 'm',
+  isMarried: true,
+  partnerName: 'Natasha',
+};
+
+export const olya = {
+  name: 'Olya',
+  sex: 'f',
+  isMarried: true,
+  partnerName: 'Maksym',
+};
+
+export const alex = {
+  name: 'Alex',
+  age: 25,
+  sex: 'm',
+  isMarried: false,
+};
